pyproject.toml
[project]
name = "stock-watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "ollama>=0.4.8",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "truthbrush",
]

[tool.uv.sources]
truthbrush = { git = "https://github.com/stanfordio/truthbrush.git" }
torchvision = [
    { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
